separate api functions and helpers (in tests)
